Moves Pydantic schemas to `schemas.py` module
